fix(daemon): share the awaiting-remote predicate through core types
The runtime reconciler imported isAwaitingRemoteExecutor from the termination
coordinator module, which the reconciler unit tests replace with a mock; the
helper resolved to undefined and every stranded remote candidate failed. Move
the pure Task predicate into @agor/core types, import it from there in both
the coordinator and the reconciler, and assert in the reconciler test that a
discovered pre-connection remote request resumes as deadline-expired.
